feat: refactor and add explain tools
