+++ .gitignore
@@ -1,3 +1,4 @@
 /venv
 .vscode/
 .env
+src/__pycache__/
